+++ query_holder.cpp
@@ -23,16 +23,16 @@ query_holder::query_holder(int dimension) {
 
     /*allocating the memory for each matrix */
 
-    host_query_B = (float*)malloc(dimension * sizeof(float)); //new double[device_num_batches * dimension];
-    host_query_A = (float*)malloc(sizeof(float)); //new double[dimension];
+    host_query_B = (float*)malloc(dimension * sizeof(float)); //new float[device_num_batches * dimension];
+    host_query_A = (float*)malloc(sizeof(float)); //new float[dimension];
 
 
     // ---------------checking host memory  for error..............................
 
     if(host_query_B == nullptr)
-        mem_error("host_query_B", "vectmatmul", dimension, "double");
+        mem_error("host_query_B", "vectmatmul", dimension, "float");
 
     if(host_query_A == nullptr)
-        mem_error("host_query_A", "vectmatmul", dimension, "double");
+        mem_error("host_query_A", "vectmatmul", dimension, "float");
 
 }
